requirements/base.txt
Django==5.1
python-environ==0.4.54
psycopg2-binary==2.9.9
django-phonenumber-field-8.0.0
phonenumbers==.13.43
djangorestframework==3.15.2
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1